Update outcomes.red
total = 21 not total == 21 
and I removed the () in the if statements because they are noise here
There also can be differences in behaviour between compiled and interpreted here and there. Especially this can be true when referencing series using deck/(i) like notation. I had some struggles with those in my Xiangqi program if I remember correctly. 
Add some smart debugging prints here and there to see if that could be the case here also.
Good luck!